Build level fleshed out
After Ryans tweaks to the level, I added in all the required pieces, tested to make sure waves worked, turrets worked, and enemies worked.
Updated Turrets and Enemies to track a list of possible targets they acquire so when one dies, it automatically switches to the next in their list (if alive)
Updated health pickups to have the health plus mesh for better visual clarity as to what it was.
Added enemies to have a drop chance to drop a health pickup.
Updated PlayerBase prefab to show a rocket ship instead of a cylinder.
Other minor tweaks/adjustments made as needed with team consent.